backend/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6